updated user panel

+++ app/Http/Controllers/UserController.php
@@ -98,7 +98,7 @@ class UserController extends Controller
     public function update(UserUpdateRequest $request, $id)
     {
         try{
-            if(auth()->user()->type != 'admin')
+            if(auth()->user()->type != 'user')
                 return $this->exceptionResponse('You are not allowed to update user',401);
 
             User::findOrFail($id)->update($request->all());
@@ -118,6 +118,9 @@ class UserController extends Controller
     public function destroy($id)
     {
         try{
+            if(auth()->user()->type != 'admin')
+                return $this->exceptionResponse('You are not allowed to update user',401);
+
             User::destroy($id);
             return $this->successResponseWithMsg('User Deleted Successfully');
         }
